feat: Split component for music contaienr
